Converted Route Calculation Code to Breadth First
The depth first route search was too slow with a 10 movement unit, testing gave 50-80ms which is frankly unacceptable. The most immediate optimization would be to convert it from a depth first to a breadth first search, this should make it find the more optimal routes to tiles within the unit's movement range faster instead of later. Performance improved on the same unit to 4-20ms, still more than 16 but at least it isn't awful.