Route Bug Fixing
Found one bug in the route calculations due to an oversight. Unit Map X/Y are anchored to their bottom-left tile. So when I check when they move one tile to the right I should check the tiles on the right edge of the unit, so that should check TargetX + unitWidth. But when I check to the left of the unit the anchor X/Y is already on the left edge so I only need to reduce the targetX by 1 instead of their width. This was causing the route code to look at the wrong tiles when it was checking movement cost when moving left or down. Hopefully this resolves a lot of the weird routing and path finding I was experiencing!