Walking on Tile Bug Fix
Found the bug with walking on tiles and inflicting walking damage on them: The code was halfway to only checking tiles the unit was walking on, half that the entire area the unit occupies deals walking damage. I settled for just the whole area for now, changing it later shouldn't be hard. This also means when the unit correctly checks the tiles they're walking on and deals appropriate damage to that tile.