Initial Ability Targeting Complete
Got ability targeting working with map points. This should mean abilities can have AoE effects based on the cursor location. I'd like to all full map targeting later but this should cover 95% of use cases.

Ability have an array of map points whicih serve as offset points from a cursor position. These offset points and the cursor postion are combined to form the target map points which then check for collision with all map objects. These map objects are pased to a targeting calculator which is responsible for checking the ability target rules and if the abilty should be used.

Next is implementing ability effects, such as a damage effect.