Abilities Doing Effects / Damage Work
Working out the implementation details of how abilities and what they do are handled.  Abilities have effects which are hardcoded in, but the parameters of these effects are set per ability. An ability that does damage has a DamageAbilityEffect attached to it, which in turn has the damage parameters such as base damage, which attributes govern the attack, and which attribute the defender gets (if any). Then work in the damage effect class to start using these parameters to start calculating damage. Gotta fill more of it in and attach it to an input state or game battle or something.

Feels like it's finally coming together!