Update to Improve Route-Clicking
Changed the order of handling inputs for the user map. It'll check for routes before checking if the user clicked on a unit. This way the user can tell a large unit to move into a spot it partially occupies, before the code thought the user told the unit to stay still and not move, making it impossible to select certain routes for units that occupy more than 1x1 squares.